issue #117 refactor: group feed component refactoring
