+++ src/pages/GroupPage/GroupPage.styles.js
@@ -6,3 +6,8 @@ export const GroupMain = styled.main`
 	margin-top: 90px;
 	padding: 0 30px;
 `;
+
+export const FeedDiv = styled.div`
+	display: flex;
+	flex-direction: column;
+`;
